resolve a -Wpointer-to-int-cast warning

+++ src/misc.c
@@ -543,7 +543,7 @@ fmd_t *fmd_create()
     md->_FileIndex = 0;
     md->KineticEnergyUpdated = false;
     md->ShowErrorMessages = true;
-    md->random_seed_aux = (int)md;
+    md->random_seed_aux = (long)md;
     _fmd_potsys_init(md);
     create_mpi_types(md);
     _fmd_h5_ds_init(md, &md->h5_dataspaces);
